+++ CHANGES.rst
@@ -1,3 +1,14 @@
+0.4.3 (unreleased)
+==================
+
+ramp_fitting
+------------
+
+- Fix issue with inappropriately including a flagged group at the beginning
+  of a ramp segment.
+
+
+
 0.4.2 (2021-10-28)
 ==================